Made an AI to select lines
Made a "derpy" level AI that just selects lines based on with box has
the least lines around it. It basically tries to lose at all costs.

Not implemented into the main game loop quite yet, simply press 'k' on
the keyboard to let the CPU take your turn.

Will be used for the quick play mode in the future and possibly a really
shitty single-player mode.